+++ pypeit/wavecalib.py
@@ -608,7 +608,7 @@ class BuildWaveCalib:
         # Inherit the calibration frame naming from self.msarc
         # TODO: Should throw an error here if these calibration frame naming
         # elements are not defined by self.msarc...
-        self.wv_calib.copy_calib_keys(self.msarc)
+        self.wv_calib.copy_calib_internals(self.msarc)
 
         # Update mask
         self.update_wvmask()