+++ CHANGES.rst
@@ -14,6 +14,7 @@
   files.  Discussed in new docs.
 - The ``'calib'`` column is now always added to the pypeit file, regardless of
   whether or not you also request the ``'comb_id'`` and ``'bkg_id'`` columns.
+- Names of associated calibration frames now written to ``spec2d`` file headers.
 
 1.12.2 (29 Mar 2023)
 --------------------